+++ CHANGELOG.md
@@ -454,7 +454,7 @@ Client Xcode projects can install the RXPromise library utilizing CocoaPods.
             return nil;
         }, ^id(NSError*error){
             for (RXPromise* p in promises) {
-                [p cancel];
+                [p cancelWithReason:error];
             }
             return error;
         });
@@ -536,4 +536,10 @@ autorelease pool.
  - Minimum Deployment target for Mac OS X is now 10.8
  
  
+### Version 0.11.3 beta (2014-04-01)
 
+ - Fixed a bug where the cancel reason for promises returned by methods `repeat` and `sequence` has not been forwarded to the current task.
+
+ - Fixed a Unit Test
+ 
+ - Added Unit Tests to confirm that cancel reasons get forwarded correctly